requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
requests==2.31.0